+++ example/app.js
@@ -1,12 +1,24 @@
 import React, { Fragment } from "react";
 import ReactDOM from "react-dom";
 
-import { Info, Success, Warning } from "../src";
+import { Icon, Info, Success, Warning } from "../src";
+import paths from "../src/icons.json";
 
 const Heading = ({ children }) => <h1>{"<"}{children}{">"}</h1>;
 
+const Icons = () => (
+  Object.keys(paths).map(icon => (
+    <span key={icon} title={icon}>
+      <Icon icon={icon} />
+    </span>
+  ))
+);
+
 const App = () => (
   <Fragment>
+    <Heading>Icon</Heading>
+    <Icons />
+
     <Heading>Info</Heading>
     <Info>This is an info.</Info>