+++ src/index.js
@@ -1,8 +1,10 @@
+import Icon from "./components/Icon";
 import Info from "./components/Info";
 import Success from "./components/Success";
 import Warning from "./components/Warning";
 
 export {
+  Icon,
   Info,
   Success,
   Warning